// cx-organization-spa/src/app/shared/constants/mass-users-creation-file-error-enum.ts
export enum MassUsersCreationFileError {
  GeneralError = 'GeneralError',
  FileIsEmpty = 'FileIsEmpty',
  FileFormatIsNotCorrect = 'FileFormatIsNotCorrect',
  FileIsNotSelected = 'FileIsNotSelected',
  FileExceedLimitRecord = 'FileExceedLimitRecord',
  FileTemplateIsInvalid = 'FileTemplateIsInvalid',
  UserRowEmptyException = 'UserRowEmptyException'
}